# rfid-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get all available users from database",
                "responses": {}
            },
            "put": {
                "description": "This endpoint updates existing user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update existing user",
                "parameters": [
                    {
                        "description": "UserBody data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserBody"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint creates a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "UserBody data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created or returned existing user",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.UserBody": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "cardID": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}